committing changes for data returns with new gateway
